refactor: move queryMemories function to dedicated operations file and add memory fetch to context
