+++ src/events/message-create/utils/relevance.ts
@@ -5,11 +5,14 @@ import { probabilitySchema, type Probability } from '@/lib/validators';
 import type { RequestHints } from '@/types';
 import { generateObject, type ModelMessage } from 'ai';
 import type { Message } from 'discord.js-selfbot-v13';
+import type { ScoredPineconeRecord } from '@pinecone-database/pinecone';
+import type { PineconeMetadataOutput } from '@/types';
 
 export async function assessRelevance(
   msg: Message,
   messages: ModelMessage[],
-  hints: RequestHints
+  hints: RequestHints,
+  memories: ScoredPineconeRecord<PineconeMetadataOutput>[]
 ): Promise<Probability> {
   try {
     const { object } = await generateObject({
